+++ Material.Blazor.MD3/Components/Checkbox/MBCheckbox.cs
@@ -15,6 +15,30 @@ public sealed class MBCheckbox : InputComponent<bool>
     #region members
 
     /// <summary>
+    /// Determines whether the button has a badge - defaults to false.
+    /// </summary>
+    [Parameter] public bool HasBadge { get; set; }
+
+    /// <summary>
+    /// The badge's style - see <see cref="MBBadgeStyle"/>, defaults to <see cref="MBBadgeStyle.ValueBearing"/>.
+    /// </summary>
+    [Parameter] public MBBadgeStyle BadgeStyle { get; set; } = MBBadgeStyle.ValueBearing;
+
+
+    /// <summary>
+    /// When true collapses the badge.
+    /// </summary>
+    [Parameter] public bool BadgeExited { get; set; }
+    private bool _cachedBadgeExited;
+
+    /// <summary>
+    /// The button's density.
+    /// </summary>
+    [Parameter] public string BadgeValue { get; set; }
+    private string _cachedBadgeValue;
+
+
+    /// <summary>
     /// Determines if the checkbox is disabled.
     /// </summary>
     [Parameter] public bool IsDisabled { get; set; } = false;