+++ Material.Blazor.MD3/Components/RadioButton/MBRadioButton.razor.cs
@@ -17,6 +17,31 @@ public partial class MBRadioButton<TItem> : InputComponent<TItem>
     #region members
 
     /// <summary>
+    /// Determines whether the button has a badge - defaults to false.
+    /// </summary>
+    [Parameter] public bool HasBadge { get; set; }
+
+    /// <summary>
+    /// The badge's style - see <see cref="MBBadgeStyle"/>, defaults to <see cref="MBBadgeStyle.ValueBearing"/>.
+    /// </summary>
+    [Parameter] public MBBadgeStyle BadgeStyle { get; set; } = MBBadgeStyle.ValueBearing;
+
+    /// <summary>
+    /// When true collapses the badge.
+    /// </summary>
+    [Parameter]
+    public bool BadgeExited { get; set; }
+    private bool _cachedBadgeExited;
+
+    /// <summary>
+    /// The button's density.
+    /// </summary>
+    [Parameter]
+    public string BadgeValue { get; set; }
+    private string _cachedBadgeValue;
+
+
+    /// <summary>
     /// The radio button's density.
     /// </summary>
     [Parameter] public MBDensity? Density { get; set; }