+++ Material.Blazor.MD3/Model.MD2/MBIconBearingSelectElement.cs
@@ -1,0 +1,13 @@
+﻿namespace Material.Blazor.MD2;
+
+/// <summary>
+/// A list item used by <see cref="MBSegmentedButtonMulti{IList{TItem}}"/>
+/// </summary>
+/// <typeparam name="T"></typeparam>
+public record MBIconBearingSelectElementMD2<T> : MBSelectElementMD2<T>
+{
+    /// <summary>
+    /// The leading icon.
+    /// </summary>
+    public string Icon { get; set; }
+}